Update Spell.java
New Spell.java class extends Thing.java and is utilized by MasterList.java.
Currently components is formatted as objects within the .json file, not strings. if we use type string this will result in a lot of reformatting...keep in mind that there are over 300 spells in the spell.json file.
spell.json file still contains a String[] tags and a String type ... which are missing from new spell class.